requirements.txt
streamlit==1.32.0
google-genai>=1.0.0
python-dotenv>=1.0.1
python-docx>=1.1.0
fpdf2>=2.7.8
pydub>=0.25.1
PyPDF2
reportlab>=3.6.12